openai-agents/examples/langfuse/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Python template using Restate with OpenAI Agent SDK"
requires-python = ">=3.12"

dependencies = [
    "hypercorn",
    "restate-sdk[serde]>=0.14.2",
    "openai-agents==0.6.5",
    "langfuse",
    "openinference-instrumentation-openai-agents"
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
]
